// emmaus_ui/src/app/services/models/create-impact-story-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateImpactStoryRequest {
  age?: number;
  beneficiaryName?: string;
  displayOrder?: number;
  fullStory?: string;
  gender?: 'FEMALE' | 'MALE';
  imageUrl?: string;
  location?: string;
  programId?: string;
  remark?: string;
  shortQuote?: string;
  status?: string;
  videoUrl?: string;
}